% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FraNchEstYN-package}
\alias{FraNchEstYN}
\alias{FraNchEstYN-package}
\title{FraNchEstYN: Crop-Disease Simulation and Calibration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to run the FraNchEstYN C# crop–disease model from R. The package prepares inputs, executes simulations or calibrations, and collects outputs in tidy data frames for further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geomodellab.github.io/FraNchEstYN}
  \item \url{https://github.com/GeoModelLab/FraNchEstYN}
  \item Report bugs at \url{https://github.com/GeoModelLab/FraNchEstYN/issues}
}

}
\author{
\strong{Maintainer}: Simone Bregaglio \email{simone.bregaglio@crea.gov.it}

Authors:
\itemize{
  \item Roberto Ferrise \email{roberto.ferrise@unifi.it}
}

}
\keyword{internal}
