% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameters_to_df}
\alias{parameters_to_df}
\title{Convert a nested parameter list back into a tidy data.frame}
\usage{
parameters_to_df(param_list)
}
\arguments{
\item{param_list}{A nested parameter list (whole model, or a single set)}
}
\value{
A tidy data.frame with columns:
  Parameter, Description, Unit, Min, Max, Value, Calibration
}
\description{
Convert a nested parameter list back into a tidy data.frame
}
